refactor(init-design-roadmap-polish): extract init-fixture helper (FINAL-S1)
Extracts the inline mkdtempSync + writeFileSync('harness.config.json') +
writeFileSync('docs/roadmap.md') scaffold that was duplicated across the matrix
test (6 scenarios) and the yes/yes e2e test into a single shared helper at
packages/cli/tests/integration/_helpers/init-fixture.ts. Rewires both consumers
to call scaffoldInitFixture({ design, roadmap }) per the spec's D7 signature.

- NEW: packages/cli/tests/integration/_helpers/init-fixture.ts. Exports
  InitFixtureScenario, InitFixtureHandle, and async scaffoldInitFixture.
  Wraps runInit, post-step-5b config mutation, and post-step-4 roadmap write.
  Returns a cleanup() callback for the test's finally block.
- MODIFIED: init-design-roadmap-matrix.test.ts. Replaces lines 89-147 inline
  scaffold with a single scaffoldInitFixture(scenario) call. All 6 scenario
  shapes preserved verbatim; every expect(...) assertion preserved byte-for-byte.
  Removed unused imports (runInit, serializeRoadmap, os, path) and local nowIso().
- MODIFIED: init-design-roadmap-yes-yes-e2e.test.ts. Replaces lines 27-79 inline
  scaffold with scaffoldInitFixture({ design: 'yes', roadmap: 'yes' }). Every
  assertion preserved byte-for-byte. Removed unused imports and local nowIso().

Notes:
- The 'Set up design system' summary string is unified to the matrix test's
  shorter form. Neither test asserts on summary content — the e2e file's
  previously-longer text was Phase 5 narrative annotation, not a behavioral
  requirement. Collapsing it loses no observable assertion.
- The helper's InitFixtureScenario interface uses 'yes' | 'no' | 'not-sure' as
  the design discriminator. Per D3 (init-design-roadmap-polish proposal),
  hyphenated 'not-sure' survives as a code-level technical identifier in
  TypeScript string-literal unions. This is outside criterion #6's grep scope
  (which targets docs/changes/init-design-roadmap-config/ and SKILL.md only);
  the Phase 4 regression-guard regex must explicitly NOT cover
  packages/cli/tests/integration/_helpers/.
- All 10 existing tests pass with zero assertion modifications (matrix 6 + e2e 1 +
  skill-catalog 3). harness validate: 290 issues, exit 1 — identical to the
  Phase 2 baseline.

Spec: docs/changes/init-design-roadmap-polish/proposal.md (FINAL-S1, D6, D7).
